nixos-asus: add nixpkgs registry pin
This ensures that commands like 'nix run nixpkgs#rtl_433' will use the same nixpkgs repo
as used for building the system. This is handy to prevent downloading all kind of deps
that were already available on the system, and to ensure packages are the same version
as which the system was built with.